Support non-integer scoring

+++ common/common/src/main/java/org/shadowrunrussia2020/android/common/models/BalanceResponse.kt
@@ -12,7 +12,7 @@ data class AccountOverview (
     var id: Int,
     var characterId: Int,
     var currentBalance: Int,
-    var currentScoring: Int,
+    var currentScoring: Float,
     var lifeStyle: String,
     var forecastLifeStyle: String
 )
